+++ ontimize/components/table/o-table.dao.ts
@@ -1,8 +1,10 @@
-import { Observable } from 'rxjs/Observable';
+import { BehaviorSubject, Observable } from 'rxjs';
 import 'rxjs/add/observable/of';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+
+import { OQueryDataArgs } from '../service.utils';
 
 export class OTableDao {
+
   usingStaticData: boolean = false;
 
   protected loadingTimer;
@@ -17,13 +19,13 @@ export class OTableDao {
   constructor(
     private dataService: any,
     private entity: string,
-    private methods: any) {
-  }
+    private methods: any
+  ) { }
 
   /**
    * Call the service query and emit data has ben modified
    */
-  getQuery(queryArgs: any): Observable<any> {
+  getQuery(queryArgs: OQueryDataArgs): Observable<any> {
     this.isLoadingResults = true;
     return this.dataService[this.methods.query].apply(this.dataService, queryArgs);
   }
@@ -50,7 +52,7 @@ export class OTableDao {
   }
 
   /**
-   * set data array and emit data has ben modified
+   * Set data array and emit data has ben modified
    * @param data
    */
   setDataArray(data: Array<any>) {
@@ -60,7 +62,7 @@ export class OTableDao {
   }
 
   setAsincronColumn(value: Array<any>, rowData: any) {
-    //Object.assign(this.data[rowIndex], value);
+    // Object.assign(this.data[rowIndex], value);
     let index = null;
     for (let i = 0; i < this.data.length; i++) {
       if (this.data[i] === rowData) {
@@ -94,4 +96,5 @@ export class OTableDao {
       clearTimeout(this.loadingTimer);
     }
   }
+
 }